Use provalloc_alloc_core to retrieve the best core
I think we should discuss that. For now I think what I did is ok, however it
seems like there is 2 part, the first one deals with core that are provisioned
to me but not allocated and the second deals with idle cores. The thing is the
fucntion we implemented on cputopology branch on linux did directly that and I
think we shouldn't separate those 2 case anymore. We should discuss about it.