Get rid off idlecore_list in schedule.c
We don't need idlecore_list anymore with the provalloc structure we built.
Here are a bunch of changes needed to remove idlecore_list from the code. For
exemple we had here to create several functions in provalloc.c that are called
in schedule.c to replace the behaviour of this idlecore_list.
Some things are not fixed yet, see the TODO.