Refactor to move alloc stuff to corealloc.c (7/8)
This commit forces the scheduler to access the process a core is either
allocated or provisioned to, via the helper functions defined in an
earlier commit. With this, there are no more direct references to *any*
fields from the sched_pcore in schedule.c anymore. There are still some
straggler references around, but motly to accomodate API calls to our
new functions in corealloc.c which require them. We should consider only
referencing cores via their pcoreid in schedule.c and removing all
references to shed_pcores from this file altogether.