Refactor to move alloc stuff to corealloc.c (8/8)
This commit moves the printing of the idlecore map from schedule.c into
corealloc.c. With this commit, all references to the idlecore map are
removed from schedule.c. There are still a few "core" specific functions
hanging around in schedule.c, but that's because they are all currently
protected by the sched_lock. We should consider either making the
locking finer grained, or externing the sched_lock so these functions
can be completely removed.