Swapped IPI sending to a full 64-bit write and APIC writes to msr regs.
Removed the shifts in all the send IPI functions in apic.h as the x2apic
uses the full 32 bit for its destination id in the ICR. We currently store
the value for the high 32 bits of the ICR and use them when we write to
msr 30. Also cleaned up iommu code and ioapic.c now calls init_irte to
remap all of its interrupts.

Also swapped IPI code to a single write to the full 64-bit register for
the ICR and to handle destinations correctly for broadcast.

Swapped all calls to apicrput and apicrget to use msr values. Removed the
definition for LAPIC_BASE as well as the check for LAPIC_BASE in
check_sym_va

Signed-off-by: GanShun <ganshun@gmail.com>