Modified IOAPIC to emit interrupts in remappable format and init irte
Modified IOAPIC code to emit remappable interrupts and to configure the
corresponding IRTE entry in the IOMMU. For simplicity's sake, the IOAPIC
IRQ numbers correspond directly to the IOMMU's IRTE index. For example,
IRQ 1 goes to IRTE index 1. Since the high portion of the IOAPIC RTEs are
no longer simply 0, some changes to IOAPICINTRINIT were made.

Signed-off-by: GanShun <ganshun@gmail.com>