Modified IOAPIC to emit interrupts in remappable format and init irte
Modified IOAPIC code to emit remappable interrupts and to configure the
corresponding IRTE entry in the IOMMU. For simplicity's sake, the IOAPIC
IRQ numbers correspond directly to the IOMMU's IRTE index. For example,
IRQ 1 goes to IRTE index 1. The source ID for the remapped interrupts
is found by traversing the DMAR structure, and looking up the BDF in the
device scope.

Signed-off-by: GanShun <ganshun@gmail.com>